requirements.txt
# Core dependencies
boto3>=1.34.0
pydantic>=2.5.0
python-dotenv>=1.0.0

# Authentication
PyJWT>=2.8.0

# Vector search and embeddings
numpy>=1.26.0
faiss-cpu>=1.7.4

# Testing
pytest>=7.4.0
pytest-cov>=4.1.0
hypothesis>=6.92.0

# Code quality
black>=23.12.0
flake8>=7.0.0
mypy>=1.8.0
bandit>=1.7.5

# NOTE: fastapi and uvicorn removed — project uses direct Lambda invocation
# NOTE: scipy removed — not required by this project
# NOTE: For production PHI enhancement: pip install presidio-analyzer presidio-anonymizer
